test post and user routes

+++ backend/src/controllers/commentController.js
@@ -1,4 +1,5 @@
 const commentServices = require('../services/commentServices');
+const postServices = require('../services/postServices');
 const catchAsync = require('../utils/catchAsync');
 
 async function createComment(req, res) {
@@ -6,11 +7,11 @@ async function createComment(req, res) {
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array() });
   }
-  const { content, userId } = req.body;
+  const { comment, userId } = req.body;
 
-  const newComment = await commentServices.createComment({ content, userId });
+  const newComment = await commentServices.createComment(comment, userId);
 
-  res.status(201).json(newComment);
+  res.status(201).json({ message: 'Comment created successfully', newComment });
 }
 
 async function getCommentsByPost(req, res) {
@@ -25,14 +26,24 @@ async function getCommentsByPost(req, res) {
 }
 
 async function deleteComment(req, res) {
-  const commentId = req.params.commentId;
-  const deletedComment = await commentServices.deleteComment(commentId);
+  const commentId = parseInt(req.params.commentId, 10);
+  const userId = req.user.id;
+  const comment = await commentServices.getCommentById(commentId);
 
-  if (!deletedComment) {
-    return res.status(404).json({ message: 'Post not found' });
+  if (!comment) {
+    return res.status(404).json({ message: 'comment not found' });
+  }
+  if (comment.userId !== userId) {
+    return res
+      .status(403)
+      .json({ error: 'You do not have permission to delete this comment' });
   }
 
-  return res.status(200).json(deletedPost);
+  await commentServices.deleteComment(commentId);
+
+  return res
+    .status(200)
+    .json({ message: 'Comment deleted successfully', comment });
 }
 
 // prob dont need a getCommentsByUser
